auto add saah, rotate more, pass args in wg scripts

+++ wireguard/del.sh
@@ -1,23 +1,17 @@
 #!/bin/bash
-# shellcheck disable=SC1091,SC2001
+# shellcheck disable=SC1091,SC2001,SC2068
 
+[ -n "$1" ] || exit 1
 pushd "$(dirname "$(readlink -f "$0")")/.." || exit 1
 source "lib.sh"
-
 peers=$(grep -- "- PEERS=" compose.yml)
 
 if grep -q "$1" <<<"$peers"; then
     sed -i "s/$peers/$(sed "s/,\?$1//" <<<"$peers")/" compose.yml
     sudo rm -rf wireguard/config/peer_"$1"
-    sudo mv -f wireguard/config/wg_confs/wg0.conf wireguard/config/wg_confs/wg0.conf.bak
-
-    if $CLS_DOCKER; then
-        sudo docker compose restart wireguard
-        sudo docker exec wireguard bash -c "wg-quick down wg0 ; wg-quick up wg0"
-        sudo docker compose up -d wireguard
-    else
-        sudo wg-quick down "$CLS_INTERN_IFACE" ; sudo wg-quick up "$CLS_INTERN_IFACE"
-    fi
+    sudo mv -f wireguard/config/wg_confs/"$CLS_INTERN_IFACE".conf wireguard/config/wg_confs/"$CLS_INTERN_IFACE".conf.bak
+    shift
+    sudo CLS_WG_ONLY=true bash restart.sh ${@@Q}
 fi
 
 popd &>/dev/null || exit
